kqueue implementation
git-svn-id: file:///source/coreland/io_poll/branches/rewrite_20070501@2225 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ iop_devpoll.c
@@ -22,7 +22,7 @@ static int iop_devp_del(struct io_poll *iop, int fd)
   return 0;
 }
 
-static long iop_devp_wait(struct io_poll *iop, int64 t)
+static int iop_devp_wait(struct io_poll *iop, int64 t)
 {
   return 0;
 }
