+++ iop_wait.c
@@ -1,7 +1,7 @@
 #include "io_poll.h"
 #include "io_poll_impl.h"
 
-long io_poll_wait(struct io_poll *iop, int64 ms)
+int io_poll_wait(struct io_poll *iop, int64 ms)
 {
   return 1;
 }
